src/pyproject.toml
[project]
name = "cloudia_app"
version = "0.0.1"

dependencies = [
	"asyncio-mqtt",
	"pydantic",
	"pyyaml",
	"ujson"
]

[build-system]
requires = [
    "setuptools >= 67.8.0",
    "setuptools_scm >= 7.1.0"
]